Split and organize ISpatial methods into dedicated classes
