+++ src/main/java/pl/szelagi/spatial/ImmutableRadius3D.java
@@ -1,23 +1,3 @@
 package pl.szelagi.spatial;
 
-public class ImmutableRadius3D<T extends Number> {
-	private final T x, y, z;
-
-	public ImmutableRadius3D(T x, T y, T z) {
-		this.x = x;
-		this.y = y;
-		this.z = z;
-	}
-
-	public T getX() {
-		return x;
-	}
-
-	public T getY() {
-		return y;
-	}
-
-	public T getZ() {
-		return z;
-	}
-}
+public record ImmutableRadius3D<T extends Number>(T x, T y, T z) {}
