+++ src/main/java/pl/szelagi/spatial/MathMethods.java
@@ -1,0 +1,9 @@
+package pl.szelagi.spatial;
+
+public class MathMethods {
+    public static boolean isBetween(double p, double a, double b) {
+        double min = Math.min(a, b);
+        double max = Math.max(a, b);
+        return p >= min && p <= max;
+    }
+}
